% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merge_ppgrid.R
\name{merge_ppgrid}
\alias{merge_ppgrid}
\title{Merge a \code{ppgrid} object to new data.}
\usage{
merge_ppgrid(id_var, time_var, ppgrid, na_fill = FALSE)
}
\arguments{
\item{id_var}{Vector identifying the groups.}

\item{time_var}{Numeric vector of time, there is no support for character
time vectors or other datetime datatypes yet.}

\item{ppgrid}{\code{ppgrid} object produced by \code{\link{train_ppgrid}}.}

\item{na_fill}{Boolean: whether to merge the furthest in the future values
from train_ppgrid in the event that the new data's times are more advanced
than the trained range.}
}
\value{
list containing the merged mean response and total weight for the
input after merging with \code{ppgrid}. Ordered as input vectors.
}
\description{
Find the just-in-time summarised response and weight for new data
as informed by a \code{ppgrid} object.
As \code{\link{train_ppgrid}}, \code{merge_ppgrid} expects ordered vectors.
}
\examples{
set.seed(1234)
ppdf <- make_longitudinal_data(1000)
ppdf_train <- ppdf[ppdf$date < 1150, ]
ppdf_test <- ppdf[ppdf$date >= 1150, ]
ppgrid <- train_ppgrid(ppdf_train$id,
                       ppdf_train$date,
                       ppdf_train$resp,
                       lag = 25,
                       window_size = 25,
                       granularity = 25)

merge_ppgrid(ppdf_test$id,
             ppdf_test$date,
             ppgrid)
}

