% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ppR.R
\docType{package}
\name{ppR}
\alias{ppR}
\alias{ppR-package}
\title{ppR: prior period variable creation}
\description{
Make prior period summary variables from longitudinal data.
}
\details{
ppR contains methods to handle the creation and implementation of by-group
weighted mean and count variables for preceding records of the same group in
data which spans time.
The main function for training these variables is \code{\link{train_ppgrid}},
which will create a just-in-time summary for each time-group combination.
This can be scored on new data with the method \code{\link{merge_ppgrid}},
which returns the just-in-time weighted mean and count for a new dataframe as
a list.

The time variable is currently expected to be strictly a numeric, with other
parameters supplied on the same scale. Support of datetime objects is
planned.
}

